requirements.txt
fastapi
asyncio
uvicorn[standard]
python-magic
requests
slowapi